Issue #20494: Update extractor to split all-examples-in-one and all-usecases-in-one generation
