+++ extractor/src/main/java/com/example/extractor/ReadmeGenerator.java
@@ -36,6 +36,21 @@ public final class ReadmeGenerator {
             "https://raw.githubusercontent.com/checkstyle/test-configs/main/";
 
     /**
+     * The default subfolder name for all-examples-in-one configs.
+     */
+    private static final String ALL_EXAMPLES_SUBFOLDER = "all-examples-in-one";
+
+    /**
+     * The default title used for all-examples-in-one READMEs.
+     */
+    private static final String ALL_EXAMPLES_TITLE = "All Examples in One Configs";
+
+    /**
+     * The title used for all-usecases-in-one READMEs.
+     */
+    private static final String ALL_USECASES_TITLE = "All UseCases in One Configs";
+
+    /**
      * Private constructor to prevent instantiation of this utility class.
      *
      * @throws UnsupportedOperationException If this constructor is invoked.
@@ -106,7 +121,8 @@ public final class ReadmeGenerator {
     }
 
     /**
-     * Generates a README file for all-in-one examples.
+     * Generates a README file for all-examples-in-one configs. Kept for backward
+     * compatibility; delegates to the generalized all-in-one README generator.
      *
      * @param allInOneFolder The folder containing all examples.
      * @param moduleName The name of the module.
@@ -117,24 +133,48 @@ public final class ReadmeGenerator {
             final Path allInOneFolder,
             final String moduleName)
             throws IOException {
-        if (allInOneFolder == null || moduleName == null) {
+        generateAllInOneReadme(allInOneFolder, moduleName,
+                ALL_EXAMPLES_SUBFOLDER, ALL_EXAMPLES_TITLE);
+    }
+
+    /**
+     * Generates a README file for an "all-in-one" folder (either all-examples-in-one
+     * or all-usecases-in-one), using the given subfolder name and title.
+     *
+     * @param allInOneFolder The folder containing all examples or use cases.
+     * @param moduleName The name of the module.
+     * @param subfolderName The subfolder name to reference in generated links
+     *                      (e.g. "all-examples-in-one" or "all-usecases-in-one").
+     * @param title The heading/title to use in the generated README
+     *              (e.g. "All Examples in One Configs" or "All UseCases in One Configs").
+     * @throws IOException If an I/O error occurs.
+     * @throws IllegalArgumentException if the argument is invalid.
+     */
+    public static void generateAllInOneReadme(
+            final Path allInOneFolder,
+            final String moduleName,
+            final String subfolderName,
+            final String title)
+            throws IOException {
+        if (allInOneFolder == null || moduleName == null
+                || subfolderName == null || title == null) {
             throw new IllegalArgumentException(
-                    "allInOneFolder and moduleName must not be null");
+                    "allInOneFolder, moduleName, subfolderName and title must not be null");
         }
 
         final String readmeContent = String.format(Locale.US,
-            "# All Examples in One Configs%n%n"
+            "# %s%n%n"
                     + "Bellow are two options that will do the same but use different versions%n"
                     + "of github actions in checkstyle repository.%n%n%n"
                     + "### Option 1%n"
                     + "Trigger report generation by comment in Pull Request:%n"
                     + "```%n"
-                    + "Github, generate report for %s/all-examples-in-one%n"
+                    + "Github, generate report for %s/%s%n"
                     + "```%n%n"
                     + "### Option 2%n%n"
                     + "Paste below given to PR description to use such test configs:%n"
                     + "```%n"
-                    + "Report label: %s/all-examples-in-one%n"
+                    + "Report label: %s/%s%n"
                     + "Diff Regression config: %s%n"
                     + "Diff Regression projects: %s%n"
                     + "```%n%n"
@@ -146,13 +186,34 @@ public final class ReadmeGenerator {
                     + "```%n"
                     + "Github, generate report for configs in PR description%n"
                     + "```%n",
-                moduleName,
-                moduleName,
-                GITHUB_RAW_URL + moduleName + "/all-examples-in-one/config.xml",
-                GITHUB_RAW_URL + moduleName + "/all-examples-in-one/list-of-projects.properties"
+                title,
+                moduleName, subfolderName,
+                moduleName, subfolderName,
+                GITHUB_RAW_URL + moduleName + "/" + subfolderName + "/config.xml",
+                GITHUB_RAW_URL + moduleName + "/" + subfolderName + "/list-of-projects.properties"
         );
 
         final Path readmePath = allInOneFolder.resolve("README.md");
         Files.writeString(readmePath, readmeContent, StandardCharsets.UTF_8);
+    }
+
+    /**
+     * Gets the standard title for a given all-in-one subfolder name.
+     *
+     * @param subfolderName The subfolder name (e.g. "all-examples-in-one"
+     *                       or "all-usecases-in-one").
+     * @return "All Examples in One Configs" for all-examples-in-one,
+     *         "All UseCases in One Configs" for anything else (including
+     *         all-usecases-in-one).
+     */
+    public static String getAllInOneTitle(final String subfolderName) {
+        final String title;
+        if (ALL_EXAMPLES_SUBFOLDER.equals(subfolderName)) {
+            title = ALL_EXAMPLES_TITLE;
+        }
+        else {
+            title = ALL_USECASES_TITLE;
+        }
+        return title;
     }
 }
